+++ lib/shaped.rb
@@ -30,6 +30,7 @@ module Shaped
       when Shaped::Shape then shape_description
       when Hash then Shaped::Shapes::Hash.new(shape_description)
       when Array then Shaped::Shapes::Array.new(shape_description)
+      when Symbol then Shaped::Shapes::Method.new(shape_description)
       when Class then Shaped::Shapes::Class.new(shape_description, validation_options)
       else
         if shape_description.respond_to?(:call)